WIP Notices

+++ notification_manager/models.py
@@ -1,5 +1,5 @@
-import cPickle
 import base64
+import cPickle
 
 from django.db import models
 
@@ -9,11 +9,24 @@ from backends.email_backend import EmailBackend
 class EnqueuedEmailNotice(models.Model):
     # This is not in the same database as users : we identify users by emails
     sent = models.BooleanField()
+    async_rt = models.BooleanField()
+    in_daily_digest = models.BooleanField()
+    in_weekly_digest = models.BooleanField()
     context = models.TextField()
     content = models.TextField()
 
     @staticmethod
-    def queue(receiver, notification_content_provider, notification_context_provider):
+    def enqueue(receiver, notification_context_provider,
+              in_rt, in_daily_digest, in_weekly_digest):
+        """
+        Enqueue an email notice in the notification database for further handling
+        :param receiver:
+        :param notification_context_provider: an instance of a class extending the
+            NotificationContextProviderMixin
+        :param in_rt: boolean, if notice should be sent asap
+        :param in_daily_digest: if notice should be included in daily digest
+        :param in_weekly_digest: if notice should be included in weekly digest
+        """
         receiver_info = {
             'receiver': {
                 'email': receiver.email,
@@ -26,7 +39,10 @@ class EnqueuedEmailNotice(models.Model):
         notification_context.update(receiver_info)
         nq_notice = EnqueuedEmailNotice(sent=False,
                                         context=base64.b64encode(cPickle.dumps(notification_context)),
-                                        content=notification_content_provider.get_content()
+                                        content=notification_context_provider.get_content(),
+                                        async_rt=in_rt,
+                                        in_daily_digest=in_daily_digest,
+                                        in_weekly_digest=in_weekly_digest
                                         )
         print nq_notice.context
 
@@ -35,7 +51,7 @@ class EnqueuedEmailNotice(models.Model):
         return nq_notice
 
     @staticmethod
-    def send_all_and_forget():
+    def send_all_immediate():
         email_backend = EmailBackend(1)
         for email_notice in EnqueuedEmailNotice.objects.all():
             context = cPickle.loads(base64.b64decode(email_notice.context.__str__()))
@@ -47,7 +63,17 @@ class EnqueuedEmailNotice(models.Model):
 
             print "Messages snt :" + str(nb_sent)
             if nb_sent > 0:
-                email_notice.delete()
+                email_notice.sent = True
+                if not (email_notice.in_daily_digest or email_notice.in_weekly_digest):
+                    email_notice.delete()
+
+    @staticmethod
+    def send_daily_digest():
+        pass
+
+    @staticmethod
+    def send_weekly_digest():
+        pass
 
 
 class EnqueuedAppNotice(models.Model):
